requirements.txt
aiogram==3.7.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pyttsx3==2.90
